Refactor the routes for flask
