+++ backend-flask/app.py
@@ -3,8 +3,6 @@ import sys
 
 from flask import Flask
 from flask import request, g
-from flask_cors import cross_origin
-from aws_xray_sdk.core import xray_recorder
 
 from lib.rollbar import init_rollbar
 from lib.xray import init_xray
@@ -13,18 +11,10 @@ from lib.cloudwatch import init_cloudwatch
 from lib.honeycomb import init_honeycomb
 from lib.helpers import model_json
 
-from services.users_short import *
-from services.home_activities import *
-from services.notifications_activities import *
-from services.user_activities import *
-from services.create_activity import *
-from services.create_reply import *
-from services.search_activities import *
-from services.message_groups import *
-from services.messages import *
-from services.create_message import *
-from services.show_activity import *
-from services.update_profile import *
+import routes.general
+import routes.activities
+import routes.users
+import routes.messages
 
 app = Flask(__name__)
 
@@ -35,179 +25,11 @@ init_cors(app)
 with app.app_context():
   g.rollbar = init_rollbar(app)
 
-@app.route('/api/health-check')
-def health_check():
-  return {'success': True, "ver": 1}, 200
-
-# Rollbar ------
-@app.route('/rollbar/test')
-def rollbar_test():
-    rollbar.report_message('Hello World!', 'warning')
-    return "Hello World!"
-
-@app.route("/api/message_groups", methods=['GET'])
-def data_message_groups():
-    access_token = extract_access_token(request.headers)
-    LOGGER.info(f"access_token:message_groups:{access_token}")
-    try:
-      claims = cognito_jwt_token.verify(access_token)
-      # authenicatied request
-      app.logger.debug("authenicated")
-      app.logger.debug(claims)
-      cognito_user_id = claims['sub']
-      model = MessageGroups.run(cognito_user_id=cognito_user_id)
-  return model_json(model)
-
-    except TokenVerifyError as e:
-      # unauthenicatied request
-      app.logger.debug(e)
-      return {}, 401
-
-@app.route("/api/messages/<string:message_group_uuid>", methods=['GET'])
-def data_messages(message_group_uuid):
-  access_token = extract_access_token(request.headers)
-  try:
-    claims = cognito_jwt_token.verify(access_token)
-    # authenicatied request
-    app.logger.debug("authenicated")
-    app.logger.debug(claims)
-    cognito_user_id = claims['sub']
-    model = Messages.run(
-        cognito_user_id=cognito_user_id,
-        message_group_uuid=message_group_uuid
-      )
-  return model_json(model)
-  except TokenVerifyError as e:
-    # unauthenicatied request
-    app.logger.debug(e)
-    return {}, 401
-
-@app.route("/api/messages", methods=['POST','OPTIONS'])
-@cross_origin()
-def data_create_message():
-  message_group_uuid   = request.json.get('message_group_uuid',None)
-  user_receiver_handle = request.json.get('handle', None)
-  message = request.json['message']
-  access_token = extract_access_token(request.headers)
-  try:
-    claims = cognito_jwt_token.verify(access_token)
-    # authenicatied request
-    app.logger.debug("authenicated")
-    app.logger.debug(claims)
-    cognito_user_id = claims['sub']
-    if message_group_uuid == None:
-      # Create for the first time
-      model = CreateMessage.run(
-        mode="create",
-        message=message,
-        cognito_user_id=cognito_user_id,
-        user_receiver_handle=user_receiver_handle
-      )
-    else:
-      # Push onto existing Message Group
-      model = CreateMessage.run(
-        mode="update",
-        message=message,
-        message_group_uuid=message_group_uuid,
-        cognito_user_id=cognito_user_id
-      )
-  return model_json(model)
-  except TokenVerifyError as e:
-    # unauthenicatied request
-    app.logger.debug(e)
-    return {}, 401
-
-@app.route("/api/activities/home", methods=['GET'])
-#@xray_recorder.capture('activities_home')
-def data_home():
-  access_token = extract_access_token(request.headers)
-  #LOGGER.info("access_token", "home", access_token)
-  LOGGER.info(f"access_token:home:{access_token}")
-  try:
-    claims = cognito_jwt_token.verify(access_token)
-    # authenicatied request
-    app.logger.debug("authenicated")
-    app.logger.debug(claims)
-    app.logger.debug(claims['username'])
-    data = HomeActivities.run(cognito_user_id=claims['username'])
-  except TokenVerifyError as e:
-    # unauthenicatied request
-    app.logger.debug(e)
-    app.logger.debug("unauthenicated")
-    data = HomeActivities.run()
-  return data, 200
-
-@app.route("/api/activities/notifications", methods=['GET'])
-def data_notifications():
-  data = NotificationsActivities.run()
-  return data, 200
-
-@app.route("/api/activities/@<string:handle>", methods=['GET'])
-#@xray_recorder.capture('activities_user')
-def data_handle(handle):
-  model = UserActivities.run(handle)
-  return model_json(model)
-
-@app.route("/api/activities/search", methods=['GET'])
-def data_search():
-  term = request.args.get('term')
-  model = SearchActivities.run(term)
-  return model_json(model)
-
-@app.route("/api/activities", methods=["POST","OPTIONS"])
-@cross_origin()
-def data_activities():
-  access_token = extract_access_token(request.headers)
-  try:
-    claims = cognito_jwt_token.verify(access_token)
-    cognito_user_id = claims['sub']
-    message = request.json['message']
-    ttl = request.json['ttl']
-    model = CreateActivity.run(message, cognito_user_id, ttl)
-  return model_json(model)
-  except TokenVerifyError as e:
-    # unauthenicatied request
-    app.logger.debug(e)
-    return {}, 401
-
-@app.route("/api/activities/<string:activity_uuid>", methods=['GET'])
-def data_show_activity(activity_uuid):
-  data = ShowActivity.run(activity_uuid=activity_uuid)
-  return data, 200
-
-@app.route("/api/activities/<string:activity_uuid>/reply", methods=['POST','OPTIONS'])
-@cross_origin()
-def data_activities_reply(activity_uuid):
-  user_handle  = 'andrewbrown'
-  message = request.json['message']
-  model = CreateReply.run(message, user_handle, activity_uuid)
-  return model_json(model)
-  return
-
-@app.route("/api/users/@<string:handle>/short", methods=['GET'])
-def data_users_short(handle):
-  data = UsersShort.run(handle)
-  return data, 200
-
-@app.route("/api/profile/update", methods=['POST','OPTIONS'])
-@cross_origin()
-def data_update_profile():
-  bio          = request.json.get('bio',None)
-  display_name = request.json.get('display_name',None)
-  access_token = extract_access_token(request.headers)
-  try:
-    claims = cognito_jwt_token.verify(access_token)
-    cognito_user_id = claims['sub']
-    model = UpdateProfile.run(
-      cognito_user_id=cognito_user_id,
-      bio=bio,
-      display_name=display_name
-    )
-  return model_json(model)
-  except TokenVerifyError as e:
-    # unauthenicatied request
-    app.logger.debug(e)
-    return {}, 401
+# load routes -----------
+routes.general.load(app)
+routes.activities.load(app)
+routes.users.load(app)
+routes.messages.load(app)
 
 if __name__ == "__main__":
   app.run(debug=True)
